requirements.txt
fastapi==0.103.2
pydantic-settings==2.0.3
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
pymysql==1.1.0
alembic==1.12.0
httpx==0.25.0
pytest==7.4.2
email-validator==2.1.0
passlib==1.7.4
bcrypt==4.0.1
fastapi-mail==1.4.1
Jinja2==3.1.2
PyJWT==2.8.0
python-multipart==0.0.6
tokenizers==0.10.1
onnxruntime
piper_phonemize
roman
langdetect
vinorm
transformers==4.10.0  
underthesea
sentencepiece
apscheduler
soundfile
torch
